mcp_rag_pdf/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3

# Database
mysql-connector-python==8.3.0
chromadb==0.4.22

# AI/ML and Embeddings
sentence-transformers==2.3.1
torch>=2.0.0
transformers>=4.30.0

# PDF Processing
PyPDF2==3.0.1
pypdf==3.17.4
pdfplumber==0.10.3
pillow==10.2.0

# Utilities
numpy==1.26.3
requests==2.31.0

# BM25 Keyword Search
rank-bm25>=0.2.2

# Neo4j Graph Database
neo4j>=5.15.0

# Entity Extraction (NER)
spacy>=3.7.0

# OpenAI for relationship extraction (already used in rag_pipeline)
# openai>=1.12.0  # Already in use